// projects/pos/src/app/shared/enums/columns/category-columns.enum.ts
export enum CategoryColumnLabel {
  NAME = 'Name',
  SLUG = 'Slug',
  SHOW_IN_MENU = 'Show In Menu',
  DISPLAY_MODE = 'Display Mode',
  STATUS = 'Status',
}

export enum CategoryColumn {
  NAME = 'name',
  SLUG = 'slug',
  SHOW_IN_MENU = 'show_in_menu',
  DISPLAY_MODE = 'display_mode',
  STATUS = 'status',
}